Updated direct UDP code on the server side to loop for continuous incoming requests
Before the server will accept one request then stop. I have tested this is working by adding a second bit of code to the client that sends a command for the user Heylon to be removed after it was added. Debug messages printed to the servers output show that this is working. I had some problems while getting it to work regarding the client using the same socket as the server when running on the same machine. May or may not need to find a way of dealing with this in future if I intend to test multiple clients on one PC