Created leader board on server
Created the LeaderBoard class on the server, which will contain all the highscores which the client requests to be posted there. Added code in the server request handler to save the scores sent to it by the client. Tested it by writing some test code on the client to request the server to save a score. The test was successful, the server responded with OKAY and print statements on the server alerted me to the successful creation of a LeaderBoard entry.