Implemented the TURN left/right handlers on the server
These requests don't return any confirmation from the server.

Although I managed to get it working, I need to make a separate function on the Client for sending requests that wont generate a response. Currently the client will block for 2 seconds waiting for a response from the server until it times out. If the request won't generate a server response, that's 2 seconds the client wastes blocking for a response that will never arrive.

Otherwise, the client was able to send a turn left, or right request to the server. I was able to see that this worked through print statements on the server that alerted me to the fact it had received that specific type of request from a client.