Implemented jetwall switch on/off request on server
The client can now send the server a request to turn the jetwall on or off.

Tested by sending such requests from the client to the server. The server was able to detirmine it was a jet wall request and printed a message to show me that it had successfully received it