added getGameState() to the client LightCyclesGame class
This will be used to query the server about the current game state. E.g. when a user wants to add themselves to the game, the client needs to detirmine if the server is waiting for players.

Tested this method by querying the servers game state from the client and printing the result. The result was a success with the client printing the correct game state