Added joinServer method to LightCyclesGame on the Client
The client can now request to join the server (Add a Player with the desired name to the Player list on the server).

Tested by calling joinServer("Heylon") on the client. I tested it with the server not running, the server running but not waiting for players, and the server running and waiting for players. joinServer threw an exception as expected when the server either wasn't running or waiting for players. The server responded with OKAY when it was waiting and added the Player to it's list