Created Player class for Client
The Player class will be used to store the player name, direction, position and possibly more in future.

The names and other values in instances of the class will be set by the server broadcast message. This class will merely be used to hold the details of each player for use between each broadcast message.