Server: Added moving speed methods on Player
The player speed on the grid can now be set. The player can be stopped, moving slowly, or moving quickly. The player is initialised with the stopped speed. The speed can only be changed to slow, or fast through the methods because once a player starts moving, it cannot be stopped again.