Server: Partially implemented progressGame method on GameGrid
This method is called every time the game needs to progress. I.E when the players need to be moved along their direction of travel whilst checking for collisions and placing jetwall objects in their wake if enabled. Bit of a mouthful, I know. This method will probably be called a set amount of times per second. I guess think of it like Frames Per Second. The more times per second this method is called, the faster, but smoother, the game will appear to run on screen.