Added code to test User Story 1
I added temporary test code to the server to quickly test if I could move and steer my light cycle around the grid (User story 1).

The test revealed that the server did respond to my requests to start moving the light cycle (Player) and to turn left and right. It moved my light cycle around the grid, at the requested speed and direction. So it was a success in that regard. There are still some unhandled conditions, such as when the light cycle hits the edge of the grid. That causes an index out of range exception on the grid 2D array. That is however a separte user story so at least for now user story 1 seems to be implemented and working.