Updated code to implement the jetwall status in the light cycles data protocol
Updated the GameState class to contain a list of PlayerStates. This
class previously contained a HashMap of player names and their
positions. This had to be changed with the addition of the jetwall
status.

Tested this working by setting up a server, and joining it with a
client, and printing the broadcast message to screen.