Created join game JPanel for the client GUI
The join game panel contains various components that allow the user to
enter their name and join the game (be placed on the grid)

The join game panel raises a custom event when the join game button is
clicked. The name that the player typed in is parsed to the
eventlisteners through the EventObject.

Tested it by displaying the join game panel in the main Jframe. I
subscribed to the custom join game event and printed the players name in
a message box when the join game button was clicked. It worked.