Created ServerRequester class for the Client
The ServerRequester class now contains the code that transmits direct
requests to the server. The code for that used to be directly on the
LightCyclesGame object. Moving it into it's own object made the code
neater.

Tested to ensure it works by setting up a server, connecting with the
client, and controlling my light cycle on the grid.