Client: Jetwall only gets added to screen if player has moved
Before hand, even if the player was stationary (I.E. when the game first
starts and no-one is moving yet) the client just keeps on spitting out
jetwalls at update after update in the same position, because the player
isn't moving. Had no real visual effect because each jetwall would
obscure the one beneath it, but this is very inefficient, as they're
never seen but still require drawing.